% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{align_chr}
\alias{align_chr}
\title{Align text on center text (default is decimal)}
\usage{
align_chr(
  x,
  accuracy = NULL,
  trim_leading_zeros = FALSE,
  add_plusses = FALSE,
  padding_character = " ",
  center = "\\\\.",
  format_integers = FALSE,
  side = c("both", "left", "right"),
  NA_value = "",
  ...
)
}
\arguments{
\item{x}{vector (numeric or character)}

\item{accuracy}{number to round to. If NULL, the current default accuracy set with \code{\link[=apa7_defaults]{apa7_defaults()}} will be used.}

\item{trim_leading_zeros}{if TRUE (default), trims leading zeros, otherwise keeps them}

\item{add_plusses}{if TRUE (default), adds a plus to positive numbers}

\item{padding_character}{character to use for padding, default is \verb{&numsp;} (figure space)}

\item{center}{text on which to align text. Center on decimal by default, but can be any text.}

\item{format_integers}{If TRUE, integers will be formatted with digits}

\item{side}{side on which to make text of equal width}

\item{NA_value}{value to replace NA}

\item{...}{additional arguments passed to \code{signs::signs()}}
}
\value{
character vector
}
\description{
Align text on center text (default is decimal)
}
\examples{
align_chr(c(1, 10, 100))
}
