% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{apa7_defaults}
\alias{apa7_defaults}
\title{Set defaults for apa7 package}
\usage{
apa7_defaults(
  accuracy = NULL,
  font_family = NULL,
  intercept_text = NULL,
  parameter_formatter = NULL,
  number_formatter = NULL,
  trim_leading_zero = NULL,
  reset = FALSE
)
}
\arguments{
\item{accuracy}{numeric (default: .01)}

\item{font_family}{font family}

\item{intercept_text}{what to call the intercept}

\item{parameter_formatter}{column formatting functions}

\item{reset}{if \code{TRUE}, reset all defaults (except as specified)}
}
\value{
previous defaults
}
\description{
Set defaults for apa7 package
}
\examples{
apa7_defaults(accuracy = .001)
# Reset to package defaults
apa7_defaults(reset = TRUE)
}
