% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_chisq}
\alias{apa_chisq}
\title{Make contingency table with chi-square test of independence}
\usage{
apa_chisq(
  x,
  suppress_warnings = TRUE,
  family = "Times New Roman",
  font_size = 12,
  line_spacing = 2,
  border_color = "black",
  border_width = 0.5
)
}
\arguments{
\item{x}{A two-column data.frame or tibble}

\item{suppress_warnings}{Suppress any warnings if true.}

\item{family}{font family}

\item{font_size}{font size}

\item{line_spacing}{spacing between lines}

\item{border_color}{border color}

\item{border_width}{border width in pixels}
}
\value{
flextable
}
\description{
Make contingency table with chi-square test of independence
}
\examples{
apa_chisq(mtcars[, c("am", "gear")])
}
