% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{apa_format_columns}
\alias{apa_format_columns}
\title{Format data columns}
\usage{
apa_format_columns(
  data,
  formatter = NULL,
  columns = NULL,
  rename_headers = TRUE,
  latex_headers = FALSE
)
}
\arguments{
\item{data}{data set (data.frame or tibble)}

\item{formatter}{\code{apa_parameter_formatter} object. If NULL, the current default formatter set with \code{\link[=apa7_defaults]{apa7_defaults()}} will be used.}

\item{columns}{(optional) vector of columns to format}

\item{rename_headers}{if \code{TRUE}, rename headers with markdown or latex}

\item{latex_headers}{if \code{TRUE}, rename headers with latex instead of markdown}
}
\value{
tibble
}
\description{
Format data columns
}
\examples{
lm(mpg ~ cyl + wt, data = mtcars) |>
  parameters::parameters() |>
  apa_format_columns() |>
  apa_flextable() |>
  colformat_md(part = "header")
}
