% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_loadings}
\alias{apa_loadings}
\title{print loadings}
\usage{
apa_loadings(
  fit,
  sort_loading = TRUE,
  min_loading = 0.2,
  column_formats = NULL,
  complexity = FALSE,
  uniqueness = FALSE
)
}
\arguments{
\item{fit}{model fit object}

\item{sort_loading}{sort table using \code{psych::fa.sort}}

\item{min_loading}{minimum loading to display}

\item{column_formats}{column_formats object to format columns. If NULL, the default column_formats is used.}

\item{complexity}{print complexity column in factor analysis table}

\item{uniqueness}{print uniqueness column in factor analysis table}
}
\value{
tibble
}
\description{
print loadings
}
