% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{apa_p_star_note}
\alias{apa_p_star_note}
\title{Make star notes for p-values}
\usage{
apa_p_star_note(x = c(0.05, 0.01, 0.001), first_alpha_marginal = FALSE)
}
\arguments{
\item{x}{vector of alpha values (p-value thresholds)}

\item{first_alpha_marginal}{if TRUE, the first alpha value is treated as marginal and gets a dagger instead of a star}
}
\value{
character vector
}
\description{
Make star notes for p-values
}
\examples{
apa_p_star_note()
apa_p_star_note(x = c(.10, .05, .01, .001), first_alpha_marginal = TRUE)
}
