% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_performance}
\alias{apa_performance}
\alias{apa_performance.lm}
\title{format model performance metrics in APA style}
\usage{
apa_performance(fit, metrics = c("R2", "Sigma"), column_formats = NULL)

\method{apa_performance}{lm}(fit, metrics = c("R2", "Sigma"), column_formats = NULL)
}
\arguments{
\item{fit}{model fit object}

\item{metrics}{performance metrics. Default is R2 and Sigma}

\item{column_formats}{column_formats object to format columns. If NULL, the default column_formats is used.}
}
\value{
tibble
}
\description{
format model performance metrics in APA style
}
\examples{
lm(mpg ~ cyl + wt, data = mtcars) |>
   apa_performance() |>
   apa_flextable()
}
