% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{hanging_indent}
\alias{hanging_indent}
\title{Return markdown text with hanging indent}
\usage{
hanging_indent(
  x,
  indent = 4,
  width = 30,
  space = NULL,
  newline = "\\\\\\\\\\n",
  whitespace_only = FALSE
)
}
\arguments{
\item{x}{text}

\item{indent}{number of spaces to indent}

\item{width}{number of characters to break lines}

\item{space}{indenting space character (defaults to non-breaking space)}

\item{newline}{text for creating new line}

\item{whitespace_only}{wrapping spaces only}
}
\value{
character vector
}
\description{
Return markdown text with hanging indent
}
\examples{
hanging_indent("Hello Darkness, my old friend. I've come to talk with you again.")
}
