% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apquarto.R
\name{install_apaquarto}
\alias{install_apaquarto}
\title{install the apaquarto extension from https://github.com/wjschne/apaquarto}
\usage{
install_apaquarto(make_project = TRUE)
}
\arguments{
\item{make_project}{After importing apaquarto to folder, make that folder a project.}
}
\description{
install the apaquarto extension from https://github.com/wjschne/apaquarto
}
\examples{
install_apaquarto
}
