% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaquarto.R
\name{make_apaquarto}
\alias{make_apaquarto}
\title{Run shiny app to make a document in APA style via Quarto}
\usage{
make_apaquarto(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{run shiny app in default browser}
}
\value{
Runs a shiny app that creates apaquarto documents
}
\description{
A wrapper for \code{shiny::runGitHub} Note that running this function will install any missing packages needed to run the app: bsicons, bslib, conflicted, dplyr, fresh, purrr, rclipboard, readr, shiny, shinyWidgets, snakecase, tibble, tidyr, tippy, toastui, yaml
}
\examples{
\dontrun{
make_apaquarto()
}
}
