% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{num_pad}
\alias{num_pad}
\title{Pads text on the left or right so that the width is the same for each element of the vector}
\usage{
num_pad(x, pad_left = TRUE, padding_character = "&numsp;", NA_value = "")
}
\arguments{
\item{x}{vector of text}

\item{pad_left}{if TRUE (default), pads on the left, otherwise pads on the right}

\item{padding_character}{character to use for padding, default is \verb{&numsp;} (figure space)}

\item{NA_value}{value to replace NA}
}
\value{
character vector
}
\description{
Pads text on the left or right so that the width is the same for each element of the vector
}
\examples{
num_pad(c("a", "bb"))
}
