% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pretty_widths}
\alias{pretty_widths}
\title{Use flextable::dim_pretty to fit column widths}
\usage{
pretty_widths(
  x,
  min_width = 0.05,
  unit = c("in", "cm", "mm"),
  table_width = 6.5
)
}
\arguments{
\item{x}{flextable}

\item{min_width}{minimum width of columns}

\item{unit}{Can be \verb{in}, \code{cm}, or \code{mm}}

\item{table_width}{width of table}
}
\value{
flextable
}
\description{
Use flextable::dim_pretty to fit column widths
}
