% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa7-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{colformat_md}
\alias{as_paragraph_md}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ftExtra}{\code{\link[ftExtra]{as_paragraph_md}}, \code{\link[ftExtra]{colformat_md}}}
}}

