% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{star_balance}
\alias{star_balance}
\title{Prefix text with figure spaces to balance star text}
\usage{
star_balance(x, star = "\\\\*", superscript = TRUE)
}
\arguments{
\item{x}{character vector}

\item{star}{star text}

\item{superscript}{Place superscript text if \code{TRUE}}
}
\value{
character vector
}
\description{
Prefix text with figure spaces to balance star text
}
\examples{

star_balance(".05\\\\^\\\\*\\\\*\\\\^")
}
