% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{str_wrap_equal}
\alias{str_wrap_equal}
\title{Like \code{stringr::str_wrap}, but attempts to create lines of equal width}
\usage{
str_wrap_equal(x, max_width = 30L, sep = "\\n")
}
\arguments{
\item{x}{character}

\item{max_width}{maximum line width}

\item{sep}{separation character}
}
\value{
character
}
\description{
Like \code{stringr::str_wrap}, but attempts to create lines of equal width
}
\examples{
str_wrap_equal("This function attempts to split the string into lines with roughly equal width.")
}
