% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{style_apa}
\alias{style_apa}
\alias{style_apa.gt_tbl}
\alias{style_apa.flextable}
\title{Style object with APA style}
\usage{
style_apa(
  x,
  family = "Times New Roman",
  font_size = 12,
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)

\method{style_apa}{gt_tbl}(
  x,
  family = "Times New Roman",
  font_size = 12,
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)

\method{style_apa}{flextable}(
  x,
  family = "Times New Roman",
  font_size = 12,
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)
}
\arguments{
\item{x}{object}

\item{family}{font family}

\item{font_size}{font size}

\item{border_color}{border color}

\item{border_width}{border width in pixels}

\item{line_spacing}{spacing between lines}
}
\value{
object
}
\description{
Style object with APA style
}
\examples{
d <- data.frame(x = 1:3, y = 4:6)
flextable::flextable(d) |>
  style_apa()
gt::gt(d) |>
  style_apa()
}
