% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{tagger}
\alias{tagger}
\alias{bold_md}
\alias{italic_md}
\alias{superscript_md}
\alias{subcript_md}
\alias{header_md}
\title{Surrounds text with tags unless empty}
\usage{
tagger(x, tag = "<span>", right_tag = gsub("^<", "</", tag))

bold_md(x)

italic_md(x)

superscript_md(x)

subcript_md(x)

header_md(x, level = 1)
}
\arguments{
\item{x}{character vector}

\item{level}{heading level}

\item{fence}{character vector}
}
\value{
character vector
}
\description{
Surrounds text with tags unless empty
}
\examples{
x <- c("hello", "", NA)
tagger(x, "<span>")
bold_md(x)
italic_md(x)
superscript_md(x)
subscript_md(x)
header_md("Level 1")
header_md("Level 2", 2)
}
