% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_star_column}
\alias{add_star_column}
\title{Add columns that separate significance stars from numbers}
\usage{
add_star_column(
  data,
  ...,
  superscript = TRUE,
  star = "\\\\*",
  star_replace = "\\\\\\\\*"
)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{column names or tidyselect function}

\item{superscript}{make stars superscript}
}
\value{
data.frame or tibble
}
\description{
Add columns that separate significance stars from numbers
}
\examples{
tibble::tibble(x = c(".45", ".58*", ".68**"),
               y = c(1,2,3),
               z = 4:6) |>
               add_star_column(x)
}
