% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa7-package.R
\docType{package}
\name{apa7-package}
\alias{apa7}
\alias{apa7-package}
\title{apa7: Write Documents in APA Style 7th Edition}
\description{
Functions to create APA Style 7th Edition documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/apa7}
  \item \url{https://wjschne.github.io/apa7/}
  \item Report bugs at \url{https://github.com/wjschne/apa7/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

}
\keyword{internal}
