% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_chisq}
\alias{apa_chisq}
\title{Make contingency table with chi-square test of independence}
\usage{
apa_chisq(
  x,
  note = NULL,
  font_family = NULL,
  font_size = 12,
  text_color = "black",
  line_spacing = 2,
  border_color = "black",
  border_width = 0.5,
  suppress_warnings = TRUE
)
}
\arguments{
\item{x}{A two-column data.frame or tibble}

\item{note}{Custom note (overrides automatic note.)}

\item{font_family}{font family}

\item{font_size}{font size}

\item{text_color}{text color}

\item{line_spacing}{spacing between lines}

\item{border_color}{border color}

\item{border_width}{border width in pixels}

\item{suppress_warnings}{Suppress any warnings if true.}
}
\value{
flextable
}
\description{
Make contingency table with chi-square test of independence
}
\examples{
apa_chisq(mtcars[, c("am", "gear")])
}
