% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_flextable}
\alias{apa_flextable}
\title{Make flextable with merged row titles according to selected column}
\usage{
apa_flextable(
  data,
  row_title_column = NULL,
  row_title_prefix = "",
  row_title_sep = " ",
  row_title_border = list(color = "gray20", style = "solid", width = 1),
  col_keys = names(data),
  cwidth = 0.75,
  cheight = 0.25,
  ...
)
}
\arguments{
\item{data}{data.frame or tibble}

\item{...}{arguments passed to flextable::flextable}

\item{column}{quoted column name to group rows}

\item{prefix}{to be added to each title}

\item{sep}{separator for prefix}
}
\value{
flextable
}
\description{
Make flextable with merged row titles according to selected column
}
\examples{
library(dplyr)
mtcars \%>\%
  select(vs, am, gear, carb) \%>\%
  pivot_longer(-vs,  names_to = "Variable") \%>\%
  dplyr::summarise(Mean = round(mean(value), 2),
                   SD = round(sd(value), 2),
                   .by = c(Variable,vs)) \%>\%
  mutate(vs = factor(vs, levels = 0:1, labels = c("Automatic", "Manual"))) \%>\%
  flextable_row_title("vs")

}
