% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_performance}
\alias{apa_performance}
\alias{apa_performance.lm}
\title{format model performance metrics in APA style}
\usage{
apa_performance(
  fit,
  performance_metrics = c("R2", "Sigma"),
  column_formatter = NULL
)

\method{apa_performance}{lm}(
  fit,
  performance_metrics = c("R2", "Sigma"),
  column_formatter = NULL
)
}
\arguments{
\item{fit}{model fit object}

\item{performance_metrics}{performance metrics. Default is R2 and Sigma}
}
\value{
tibble
}
\description{
format model performance metrics in APA style
}
\examples{
lm(mpg ~ cyl + wt, data = mtcars) |>
   apa_performance() |>
   apa_flextable() |>
   ftExtra::colformat_md(part = "all")
}
