% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_performance_comparison}
\alias{apa_performance_comparison}
\title{format model comparrison metrics in APA style}
\usage{
apa_performance_comparison(
  ...,
  performance_metrics = c("R2", "deltaR2", "F", "p"),
  starred = NA,
  column_formatter = NULL
)
}
\arguments{
\item{performance_metrics}{performance metrics. Default is deltaR2}

\item{fit1}{model fit object}

\item{fit2}{modle fit object}
}
\value{
tibble
}
\description{
format model comparrison metrics in APA style
}
\examples{
m1 <- lm(mpg ~ cyl, data = mtcars)
m2 <- lm(mpg ~ cyl + wt, data = mtcars)
apa_performance_comparison(list(`Model 1` =m1, `Model 3` =m2)) |>
   apa_flextable() |>
   colformat_md(part = "all")
}
