% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{apa_style}
\alias{apa_style}
\alias{apa_style.gt_tbl}
\alias{apa_style.flextable}
\title{Style object with APA style}
\usage{
apa_style(
  x,
  family = "Times New Roman",
  font_size = 12,
  text_color = "black",
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)

\method{apa_style}{gt_tbl}(
  x,
  family = "Times New Roman",
  font_size = 12,
  text_color = "black",
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)

\method{apa_style}{flextable}(
  x,
  family = "Times New Roman",
  font_size = 12,
  text_color = "black",
  border_color = "black",
  border_width = 0.5,
  line_spacing = 2
)
}
\arguments{
\item{x}{object}

\item{family}{font family}

\item{font_size}{font size}

\item{text_color}{text color}

\item{border_color}{border color}

\item{border_width}{border width in pixels}

\item{line_spacing}{spacing between lines}
}
\value{
object
}
\description{
Style object with APA style
}
\examples{
d <- data.frame(x = 1:3, y = 4:6)
flextable::flextable(d) |>
  apa_style()
gt::gt(d) |>
  apa_style()
}
