% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaquarto.R
\name{install_apaquarto}
\alias{install_apaquarto}
\title{Installs the apaquarto extension.}
\usage{
install_apaquarto(no_prompt = FALSE, quiet = FALSE, quarto_args = NULL)
}
\arguments{
\item{no_prompt}{Do not prompt to confirm approval to download external extension.}

\item{quiet}{Suppress warning and other messages}

\item{quarto_args}{Character vector of other quarto CLI arguments to append to the Quarto command executed by this function.}
}
\value{
installs the apaquarto Quarto extension
}
\description{
A wrapper for \code{quarto::quarto_add_extension}
}
\examples{
\dontrun{
install_apaquarto()
}
}
