% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaquarto.R
\name{make_apaquarto}
\alias{make_apaquarto}
\title{Run shiny app to make a document in APA style via Quarto}
\usage{
make_apaquarto(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{run shiny app in default browser}
}
\value{
Runs a shiny app that creates apaquarto documents
}
\description{
A wrapper for \code{shiny::runGitHub}
}
\examples{
\dontrun{
make_apaquarto()
}
}
