% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{parameter_formatter}
\alias{parameter_formatter}
\title{Create a set of column formats}
\usage{
parameter_formatter(
  .data = NULL,
  accuracy = NULL,
  intercept_text = NULL,
  starred_columns = character(0),
  variable_labels = character(0)
)
}
\arguments{
\item{accuracy}{numeric (passed to scales::number)}

\item{intercept_text}{describe intercept}

\item{starred_columns}{which columns get p-value stars}

\item{variable_labels}{named vector of variable names (with vector names as labels). For example, c(\verb{Parental Income} = "parental_income", \verb{Number of Siblings} = "n_siblings")}
}
\value{
parameter_formatter
}
\description{
Create a set of column formats
}
\examples{
my_formatter <- parameter_formatter()
my_formatter@Coefficient@formatter <- \(x) round(x, 2)
my_formatter@Coefficient@formatter(2.214)
}
